+++ src/main/java/insane96mcp/survivalreimagined/module/experience/BlockExperience.java
@@ -1,37 +1,67 @@
 package insane96mcp.survivalreimagined.module.experience;
 
-import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
+import insane96mcp.survivalreimagined.base.SRFeature;
+import insane96mcp.survivalreimagined.data.IdTagRange;
 import insane96mcp.survivalreimagined.data.generator.SRBlockTagsProvider;
 import insane96mcp.survivalreimagined.module.Modules;
 import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.block.Block;
 import net.minecraftforge.event.level.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@Label(name = "Experience From Blocks", description = "Decrease / Increase experience dropped by blocks broken")
-@LoadFeature(module = Modules.Ids.EXPERIENCE, enabledByDefault = false)
-public class BlockExperience extends Feature {
+import java.util.ArrayList;
+import java.util.List;
+
+@Label(name = "Experience from Blocks", description = "Decrease / Increase experience dropped by blocks broken. Custom xp for blocks can be added in Experience/Experience from Blocks/blocks_experience.json")
+@LoadFeature(module = Modules.Ids.EXPERIENCE)
+public class BlockExperience extends SRFeature {
 	public static final TagKey<Block> NO_BLOCK_XP_MULTIPLIER = SRBlockTagsProvider.create("no_block_xp_multiplier");
 
 	@Config(min = 0d, max = 128d)
 	@Label(name = "Experience from Blocks Multiplier", description = "Experience dropped by blocks (Ores and Spawners) will be multiplied by this multiplier. Experience dropped by blocks are still affected by 'Global Experience Multiplier'\nCan be set to 0 to make blocks drop no experience")
 	public static Double blockMultiplier = 1d;
 
+	public static final ArrayList<IdTagRange> CUSTOM_BLOCKS_EXPERIENCE_DEFAULT = new ArrayList<>(List.of(
+			new IdTagRange(IdTagRange.Type.TAG, "survivalreimagined:copper_ores", 0, 1),
+			new IdTagRange(IdTagRange.Type.TAG, "survivalreimagined:iron_ores", 0, 2),
+			new IdTagRange(IdTagRange.Type.TAG, "survivalreimagined:gold_ores", 1, 3)
+	));
+
+	public static final ArrayList<IdTagRange> customBlocksExperience = new ArrayList<>();
+
 	public BlockExperience(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
+		JSON_CONFIGS.add(new JsonConfig<>("blocks_experience.json", customBlocksExperience, CUSTOM_BLOCKS_EXPERIENCE_DEFAULT, IdTagRange.LIST_TYPE));
 	}
 
 	@SubscribeEvent
 	public void onBlockXPDrop(BlockEvent.BreakEvent event) {
 		if (!this.isEnabled()
-				|| blockMultiplier == 1.0d
-				|| !event.getState().is(NO_BLOCK_XP_MULTIPLIER))
+				|| event.getState().is(NO_BLOCK_XP_MULTIPLIER))
 			return;
 
+		handleBlockDrop(event);
+		handleMultiplier(event);
+	}
+
+	private static void handleBlockDrop(BlockEvent.BreakEvent event) {
+		int silkTouchLevel = event.getPlayer().getMainHandItem().getEnchantmentLevel(Enchantments.SILK_TOUCH);
+		if (silkTouchLevel > 0)
+			return;
+		for (IdTagRange idTagRange : customBlocksExperience) {
+			if (idTagRange.matchesBlock(event.getState().getBlock()))
+				event.setExpToDrop(idTagRange.getRandomIntBetween(event.getLevel().getRandom()));
+		}
+	}
+
+	private static void handleMultiplier(BlockEvent.BreakEvent event) {
+		if (blockMultiplier == 1d)
+			return;
 		int xpToDrop = event.getExpToDrop();
 		xpToDrop *= blockMultiplier;
 		event.setExpToDrop(xpToDrop);